CTX-5783: Mypy Linter Action

+++ tasks/bio-read-quality/main.py
@@ -9,7 +9,7 @@ import matplotlib.pyplot as plt
 from coretex import currentTaskRun, SequenceDataset, folder_manager, TaskRun
 
 
-def calculateAverageScores(qualityScores: list[list[int]]) -> list[float]:
+def calculateAverageScores(qualityScores: list[list[int]]) -> list[int]:
     maxLength = max(len(readScores) for readScores in qualityScores)
 
     totalScores = [0] * maxLength
@@ -17,19 +17,19 @@ def calculateAverageScores(qualityScores: list[list[int]]) -> list[float]:
         for i in range(len(readScores)):
             totalScores[i] += readScores[i]
 
-    return [score / len(qualityScores) for score in totalScores]
+    return [int(score / len(qualityScores)) for score in totalScores]
 
 
-def analyzeFastq(sequencePath: Path) -> list[float]:
+def analyzeFastq(sequencePath: Path) -> list[int]:
     qualityScores: list[list[int]] = []
     with sequencePath.open("r") as file:
-        for record in SeqIO.parse(file, "fastq"):
+        for record in SeqIO.parse(file, "fastq"):  # type: ignore[no-untyped-call]
             qualityScores.append(record.letter_annotations["phred_quality"])
 
     return calculateAverageScores(qualityScores)
 
 
-def createPlot(scores: list[float], title: str, plotPath: Path) -> Path:
+def createPlot(scores: list[int], title: str, plotPath: Path) -> Path:
     fig, ax = plt.subplots(figsize = (10, 6))
     ax.plot(range(len(scores)), scores, linestyle = "-", color = "b", linewidth = 2, label = "Phred Scores")
 
@@ -52,8 +52,8 @@ def main() -> None:
     taskRun.setDatasetType(SequenceDataset)
     taskRun.dataset.download()
 
-    forwardScores: list[list[float]] = []
-    reverseScores: list[list[float]] = []
+    forwardScores: list[list[int]] = []
+    reverseScores: list[list[int]] = []
 
     for sample in taskRun.dataset.samples:
         logging.info(f">> [Quality Scores] Analysing sample \"{sample.name}\"")
